move particleposition update to module
